requirements.txt
torch
transformers
gradio
easyocr
opencv-python-headless
imutils
Pillow
numpy